#55 Some refactor
