--- links-slackbot/src/main/java/juja/microservices/links/slackbot/model/SaveLinkRequest.java
@@ -1,15 +1,0 @@
-package juja.microservices.links.slackbot.model;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-/**
- * @author Ivan Shapovalov
- */
-@AllArgsConstructor
-@Getter
-public class SaveLinkRequest {
-
-    private final String url;
-
-}
