+++ links-slackbot/src/main/java/juja/microservices/links/slackbot/model/links/Link.java
@@ -1,4 +1,4 @@
-package juja.microservices.links.slackbot.model;
+package juja.microservices.links.slackbot.model.links;
 
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
