+++ links-slackbot/src/test/resources/application.properties
@@ -1,1 +1,2 @@
+spring.profiles.active=dev
 slack.slashCommandToken=slashCommandToken
